# docs/swagger.yaml
basePath: /transactions
definitions:
  apimanager.GenericResponse-apimanager_TransactionsGetResponse:
    properties:
      error:
        type: string
      result:
        $ref: '#/definitions/apimanager.TransactionsGetResponse'
    type: object
  apimanager.GenericResponse-array_apimanager_TransactionSaveResponse:
    properties:
      error:
        type: string
      result:
        items:
          $ref: '#/definitions/apimanager.TransactionSaveResponse'
        type: array
    type: object
  apimanager.GenericResponse-string:
    properties:
      error:
        type: string
      result:
        type: string
    type: object
  apimanager.Paging:
    properties:
      next_page:
        type: integer
      page:
        type: integer
    type: object
  apimanager.Transaction:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      operation:
        type: string
      origin:
        type: string
      user_id:
        type: string
    type: object
  apimanager.TransactionSaveRequest:
    properties:
      amount:
        type: number
      description:
        type: string
      operation:
        type: string
    type: object
  apimanager.TransactionSaveResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      transaction:
        $ref: '#/definitions/apimanager.Transaction'
    type: object
  apimanager.TransactionsGetRequest:
    properties:
      _id:
        type: string
      amount_greater:
        type: number
      amount_less:
        type: number
      date_from:
        type: string
      date_to:
        type: string
      description:
        type: string
      operation_type:
        type: string
      origin:
        type: string
      paging:
        $ref: '#/definitions/apimanager.Paging'
      user_id:
        type: string
    type: object
  apimanager.TransactionsGetResponse:
    properties:
      paging:
        $ref: '#/definitions/apimanager.Paging'
      transactions:
        items:
          $ref: '#/definitions/apimanager.Transaction'
        type: array
    type: object
  apimanager.TransactionsSaveRequest:
    properties:
      origin_channel:
        type: string
      transactions:
        items:
          $ref: '#/definitions/apimanager.TransactionSaveRequest'
        type: array
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: law.matsuyama@outlook.com
    name: Lawrence Matsuyama
  description: API to save and list user transactions.
  termsOfService: http://swagger.io/terms/
  title: Swagger Transactions API
  version: "2.0"
paths:
  /v1/get:
    post:
      consumes:
      - application/json
      description: List transactions by giving filter
      parameters:
      - description: Transactions Get Request
        in: body
        name: transactions_get_request
        required: true
        schema:
          $ref: '#/definitions/apimanager.TransactionsGetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimanager.GenericResponse-apimanager_TransactionsGetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimanager.GenericResponse-string'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimanager.GenericResponse-string'
      summary: API to get transactions in the application.
      tags:
      - transaction
  /v1/save:
    post:
      consumes:
      - application/json
      description: Receives transactions data, registed it in application and finish
        notifying other applications.
      parameters:
      - description: Transactions Save Request
        in: body
        name: transactions_save_request
        required: true
        schema:
          $ref: '#/definitions/apimanager.TransactionsSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimanager.GenericResponse-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimanager.GenericResponse-array_apimanager_TransactionSaveResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimanager.GenericResponse-array_apimanager_TransactionSaveResponse'
      summary: API to save transactions in the application.
      tags:
      - transaction
swagger: "2.0"